gh-65 Move all security resources
