+++ src/index.ts
@@ -27,15 +27,13 @@ export * from './federation/index';
 export * from './profiles/index';
 export * from './terminologies/index';
 export * from './reference-data/index';
+export * from './security/index';
 export * from './mdm-merge-types.model';
 export * from './mdm-classifier.model';
 export * from './mdm-classifier.resource';
 export * from './mdm-resources-configuration';
 export * from './mdm-annotation.resource';
-export * from './mdm-securable.resource';
 export * from './mdm-facets.resource';
-export * from './mdm-user-groups.resource';
-export * from './mdm-group-role.resource';
 export * from './mdm-provider.model';
 export * from './mdm-provider.resource';
 export * from './mdm-catalogue-item.model';
@@ -44,10 +42,7 @@ export * from './mdm-version-link.resource';
 export * from './mdm-admin.model';
 export * from './mdm-admin.resource';
 export * from './mdm-rest-handler';
-export * from './mdm-security.model';
-export * from './mdm-security.resource';
 export * from './mdm-namespaces.resource';
-export * from './mdm-catalogue-user.resource';
 export * from './mdm-session.model';
 export * from './mdm-session.resource';
 export * from './mdm-import-export.model';
@@ -57,7 +52,6 @@ export * from './mdm-folder.resource';
 export * from './mdm-tree-item.model';
 export * from './mdm-tree-item.resource';
 export * from './mdm-metadata.resource';
-export * from './mdm-authentication.resource';
 export * from './mdm-user-image-file.resource';
 export * from './mdm-edit.resource';
 export * from './mdm-resource';
@@ -67,10 +61,6 @@ export * from './mdm-api-properties.model';
 export * from './mdm-api-properties.resource';
 export * from './mdm-versioned-folder.model';
 export * from './mdm-versioned-folder.resource';
-export * from './mdm-plugin-openid-connect.model';
-export * from './mdm-plugin-openid-connect.resource';
-export * from './mdm-plugin-doi.model';
-export * from './mdm-plugin-doi.resource';
 export * from './mdm-merge.model';
 export * from './mdm-merge.resource';
 export * from './mdm-async-jobs.model';
