+++ src/profiles/index.ts
@@ -1,2 +1,4 @@
 export * from './mdm-profile.model';
 export * from './mdm-profile.resource';
+export * from './mdm-plugin-doi.model';
+export * from './mdm-plugin-doi.resource';
